Uploading Progress on adding JWST images
I've been working on new sub tabs in the Hubble views tab (to be renamed) that will contain the Hubble image, and another that will contain the JWST images. I have all the back end additions written, I'm just working on the GUI changes. I'll be changing the format to match the news articles, and will now be using the newslistbuilder() function to build the hubble/JWST objects as well.